backend/requirements.txt
absl-py==0.10.0
asgiref==3.2.10
astunparse==1.6.3
blis==0.4.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.1.1
cymem==2.0.3
defusedxml==0.6.0
Django==3.1.1
django-cors-headers==3.5.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
gast==0.3.3
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
gunicorn==20.1.0
h5py==2.10.0
httpie==2.3.0
idna==2.10
jdcal==1.4.1
joblib==0.16.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.2.2
murmurhash==1.0.2
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
openpyxl==3.0.5
opt-einsum==3.3.0
pandas==1.1.1
plac==1.1.3
preshed==3.0.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
PySocks==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
six==1.15.0
sklearn==0.0
spacy==2.3.2
sqlparse==0.3.1
srsly==1.0.2
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
thinc==7.4.1
threadpoolctl==2.1.0
tqdm==4.48.2
urllib3==1.25.10
wasabi==0.7.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
